pyproject.toml
[project]
name = "cosmosdb-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.5.0",
    "psycopg2-binary>=2.9.10",
]
